refactor: Lnf에서 관리자 기능 Admin으로 적재
