+++ backend/src/main/java/com/passtival/backend/domain/admin/repository/AdminRepository.java
@@ -1,11 +1,11 @@
-package com.passtival.backend.domain.lostfound.repository;
+package com.passtival.backend.domain.admin.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.passtival.backend.domain.lostfound.model.entity.Admin;
+import com.passtival.backend.domain.admin.model.entity.Admin;
 
 @Repository
 public interface AdminRepository extends JpaRepository<Admin, Long> {
